Separate Element class from DimensionSubset.
