+++ include/DimensionSubset.h
@@ -1,8 +1,9 @@
 #pragma once
 
+#include "Element.h"
+
 #include <engines/gis/CRSRegistry.h>
 #include <engines/querydata/Producer.h>
-#include <spine/Value.h>
 #include <xercesc/dom/DOMDocument.hpp>
 
 namespace SmartMet
@@ -11,22 +12,6 @@ namespace Plugin
 {
 namespace WCS
 {
-class Element
-{
- public:
-  using NameType = std::string;
-  using Value = Spine::Value;
-  Element(const xercesc::DOMElement* element);
-  Element(const Value& value, const NameType& name);
-  virtual ~Element() {}
-  const Value& getValue() const { return mValue; }
-  const NameType& getName() const { return mName; }
-
- private:
-  Value mValue;
-  NameType mName;
-};
-
 class DimensionSubset
 {
  public:
